Add support for a unified Xen binary.
In efforts for XCP-ng to support a shim-less Secure Boot boot chain,
this commit adds support into Xen for what is called a unified EFI
binary. It builds an EFI binary and teaches Xen how to extract the
kernel/initrd/config from the xen.efi image when it is launched as an
EFI payload AND contains embedded binaries. Without embedded binaries,
the EFI image launches normally.

This PR also adds the EFI image as an installed executable. In order to
preserve precious space on systems with small ESPs, it installs it *off*
of the ESP. This assumes that the EFI binary won't be used outside of
the unified binary. The unified binary (another RPM), does install the
unified EFI onto the ESP, as required for it to be bootable.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>