Support Xen as a unified EFI binary
This commit adds support for the "unified binary" technique of booting.

The requirements are two-fold:
    1. Build Xen as an EFI binary
    2. Teach Xen to find the kernel embedded in the Xen binary
       (introspecting into itself).

Number 1 is done simply by building Xen with a `binutils` package that
supports PE/COFF binaries (i386pe / i386pep).

Number 2 is accomplished by the applied patches, which introduce the
introspecting functionality. The patch that provides that capability is
unfortunately dependent on an updated version of the EFI build system,
and so many of the accompanying patches only affect the build.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>